+++ lib/main.dart
@@ -1,5 +1,4 @@
 import 'package:algolia/algolia.dart';
-import 'package:avzag/navigation/root_guard.dart';
 import 'package:avzag/store.dart';
 import 'package:firebase_core/firebase_core.dart';
 import 'package:flutter/material.dart';
@@ -8,7 +7,8 @@ import 'package:shared_preferences/shared_preferences.dart';
 import 'package:url_strategy/url_strategy.dart';
 
 import 'firebase_options.dart';
-import 'navigation/router.gr.dart';
+import 'modules/navigation/services/root_guard.dart';
+import 'modules/navigation/services/router.gr.dart';
 import 'theme_set.dart';
 
 void main() async {
