+++ lib/modules/dictionary/screens/word_loader.dart
@@ -2,8 +2,8 @@ import 'package:auto_route/auto_route.dart';
 import 'package:avzag/modules/dictionary/models/word.dart';
 import 'package:avzag/modules/dictionary/screens/word.dart';
 import 'package:avzag/modules/dictionary/services/word.dart';
-import 'package:avzag/navigation/loading_page.dart';
-import 'package:avzag/navigation/router.gr.dart';
+import 'package:avzag/modules/navigation/screens/loader.dart';
+import 'package:avzag/modules/navigation/services/router.gr.dart';
 import 'package:flutter/material.dart';
 
 class WordLoaderScreen extends StatelessWidget {
@@ -16,7 +16,7 @@ class WordLoaderScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return LoadingPage<Word>(
+    return LoaderScreen<Word>(
       loadWord(id),
       then: (context, w) async {
         if (w == null) {
