+++ lib/modules/home/screens/home.dart
@@ -1,6 +1,6 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:avzag/models/language.dart';
-import 'package:avzag/navigation/router.gr.dart';
+import 'package:avzag/modules/navigation/services/router.gr.dart';
 import 'package:avzag/shared/extensions.dart';
 import 'package:avzag/shared/modals/loading_dialog.dart';
 import 'package:avzag/shared/widgets/language_avatar.dart';
